Configure USB NIC with NetworkManager
A new qdb subcommand "network" takes a MAC address and configures the
network device having that MAC to use link-local IPv4 addresses.

If the device is already using link-local IPv4 addresses, nothing is
done. If not, the existing connections for that device are checked to
see if one of them uses link-local addresses. If a suitable connection
is found, that connection is activated. Otherwise a new connection is
created and activated.

Task-number: QTBUG-55432
Change-Id: I689eb016bffb45c0cb7999b109febd5b34b1e0e4
Reviewed-by: Kimmo Ollila <kimmo.ollila@theqtcompany.com>