Handle multiple USB devices
Add a new qdb subcommand "devices". It lists the USB devices that are
available and have a QDB interface. Their serial numbers are retrieved
from the USB interface description.

All the other commands now use --device option to decide which device to
connect to.

libusb_context is managed in a separate singleton, because managing its
lifetime between the DeviceManager and UsbConnection would be awkward
otherwise. The library-wide initialization is anyway meant to happen
just once and the deinitialization just before quitting the executable.

Task-number: QTBUG-54405
Change-Id: I8dd4f759e2e0906e0b9d121eb9f1221ee4de55ca
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>