Handle plugged and unplugged USB devices
Looking for inserted and removed devices is done by polling once a
second. libusb hotplug events are not used since they are not supported
on Windows and not available in the libusb version in RHEL 7.2.

If the IP address is not available for the device on the first check,
the device will be checked again after a short delay until the
information is complete. In the meanwhile the incomplete information is
exposed to the client when asked.

Parts that use libusb were moved from libqdb to the host part, since
they are only used on the host.

Task-number: QTBUG-56067
Change-Id: I0cfbd3659001982609a7d4a77cb1266a7dc6ce8c
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>