Strip non-ASCII characters from USB serial numbers
This works around getting the correct serial number prepended with
U+033A (COMBINING INVERTED BRIDGE BELOW). That character does not seem
to be present when the serial is set up in the qdbd-init.sh script and
it is not read by HandshakeExecutor either.

This assumes serial numbers shouldn't have non-ASCII and don't have
question marks.

Change-Id: I133a76020b2ea42a12338485a45fce2679895845
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>