Reset network configuration when cable is unplugged
Read USB function filesystem events from the control endpoint and react
to events related to disconnecting by calling the network configuration
script to reset.

This means that when the cable is plugged in, network is always in a
reset state and a stale IP according to the previous configuration won't
be returned.

Task-number: QTBUG-59450
Change-Id: I4ae77c6a9283941802d7b04b3588dbbb81b7c266
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>